fix(layering): remove foundation->horizontal/modality edges from proximadb-distance-kernel (#207)
The Validate Workspace Layering check was red on develop because the
foundation-tier proximadb-distance-kernel depended upward on:
  - proximadb-runtime-common (horizontal)
  - proximadb-vector         (modality)

Edge 1 (runtime-common): extract the pool module into a new foundation
crate proximadb-memory-pool (deps parking_lot + tracing only). Repoint
distance-kernel::engine to proximadb_memory_pool. runtime-common's pool
module is now a pub-use re-export of proximadb-memory-pool so all
existing proximadb_runtime_common::pool::* consumers are unaffected.

Edge 2 (vector): the three proximadb-vector references were inverted
re-export shims (conversion.rs, int8_simd.rs, platform/avx512.rs) with
zero consumers anywhere in the workspace. Delete the shims and the
proximadb-vector dependency. No implementation deleted; the real impls
remain in proximadb-vector::distance. The experimental-turboquant
feature is retained as a foundation-local gate (TurboQuantVectorData is
self-contained in quantized.rs, no proximadb-vector dep needed).
