--- crates/foundation/proximadb-distance-kernel/src/conversion.rs
@@ -1,7 +1,0 @@
-//! Distance metric conversion utilities
-//!
-//! **MIGRATION NOTICE**: Implementation has moved to the vector modality crate.
-//!
-//! `crates/modalities/proximadb-vector/src/distance/conversion.rs`
-
-pub use proximadb_vector::distance::conversion::*;
