--- crates/foundation/proximadb-distance-kernel/src/int8_simd.rs
@@ -1,7 +1,0 @@
-//! INT8 SIMD Distance Computation
-//!
-//! **MIGRATION NOTICE**: Implementation has moved to the vector modality crate.
-//!
-//! `crates/modalities/proximadb-vector/src/distance/int8_simd.rs`
-
-pub use proximadb_vector::distance::int8_simd::*;
