--- crates/foundation/proximadb-distance-kernel/src/platform/avx512.rs
@@ -1,1 +1,0 @@
-pub use proximadb_vector::distance::avx512::*;
