+++ crates/foundation/proximadb-memory-pool/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "proximadb-memory-pool"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "ProximaDB foundation-tier memory pooling infrastructure (reusable buffer pools to reduce allocation overhead in vector operations)"
+
+[lib]
+name = "proximadb_memory_pool"
+path = "src/lib.rs"
+
+[dependencies]
+parking_lot.workspace = true
+tracing.workspace = true
