+++ crates/horizontal/proximadb-runtime-common/Cargo.toml
@@ -17,3 +17,4 @@ parking_lot.workspace = true
 serde.workspace = true
 tokio.workspace = true
 tracing.workspace = true
+proximadb-memory-pool.workspace = true
